requirements.txt
certifi==2022.6.15
Flask==2.1.2
Flask_Cors==3.0.10
gevent==21.12.0
pymongo==4.2.0
python-dotenv==0.20.0